+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
 -----
 
 * Fixed empty :class:`.Table` column names would be intialized as list instead of tuple.
-* Added :meth:`.Table.fuse` to join tables sequentially without filtering.
+* :meth:`.Table.join` can now join by row numbers—a sequential join.
 * :meth:`.Table.join` now supports full outer joins via the ``full_outer`` keyword.
 * :meth:`.Table.join` can now accept column indicies instead of column names.
 * :meth:`.Table.from_csv` now buffers input files to prevent issues with using STDIN as an input.